+++ spec/rubocop/formatter/junit_formatter_spec.rb
@@ -6,8 +6,8 @@ RSpec.describe RuboCop::Formatter::JUnitFormatter do
   let(:output) { StringIO.new }
 
   describe '#file_finished' do
-    it 'displays parsable text' do
-      cop = RuboCop::Cop::Cop.new
+    before do
+      cop = RuboCop::Cop::Layout::SpaceInsideBlockBraces.new
       source_buffer = Parser::Source::Buffer.new('test', 1)
       source_buffer.source = %w[foo bar baz].join("\n")
 
@@ -26,29 +26,52 @@ RSpec.describe RuboCop::Formatter::JUnitFormatter do
       formatter.file_finished('test_2', cop.offenses)
 
       formatter.finished(nil)
+    end
 
-      expect(output.string).to eq(<<~XML.chop)
+    it 'displays start of parsable text' do
+      expect(output.string).to start_with(<<~XML)
         <?xml version='1.0'?>
         <testsuites>
           <testsuite name='rubocop'>
-            <testcase classname='test_1' name='Cop/Cop'>
-              <failure type='Cop/Cop' message='message 1'>
-                test:1:1
-              </failure>
-              <failure type='Cop/Cop' message='message 2'>
-                test:3:2
-              </failure>
-            </testcase>
-            <testcase classname='test_2' name='Cop/Cop'>
-              <failure type='Cop/Cop' message='message 1'>
-                test:1:1
-              </failure>
-              <failure type='Cop/Cop' message='message 2'>
-                test:3:2
-              </failure>
-            </testcase>
+      XML
+    end
+
+    it 'displays end of parsable text' do
+      expect(output.string).to end_with(<<~XML.chop)
           </testsuite>
         </testsuites>
+      XML
+    end
+
+    it "displays an offfense for `classname='test_1` in parsable text" do
+      expect(output.string).to include(<<-XML)
+    <testcase classname='test_1' name='Layout/SpaceInsideBlockBraces'>
+      <failure type='Layout/SpaceInsideBlockBraces' message='message 1'>
+        test:1:1
+      </failure>
+      <failure type='Layout/SpaceInsideBlockBraces' message='message 2'>
+        test:3:2
+      </failure>
+    </testcase>
+      XML
+    end
+
+    it "displays an offfense for `classname='test_2` in parsable text" do
+      expect(output.string).to include(<<-XML)
+    <testcase classname='test_2' name='Layout/SpaceInsideBlockBraces'>
+      <failure type='Layout/SpaceInsideBlockBraces' message='message 1'>
+        test:1:1
+      </failure>
+      <failure type='Layout/SpaceInsideBlockBraces' message='message 2'>
+        test:3:2
+      </failure>
+    </testcase>
+      XML
+    end
+
+    it 'displays a non-offfense element in parsable text' do
+      expect(output.string).to include(<<~XML)
+        <testcase classname='test_1' name='Style/Alias'/>
       XML
     end
   end